+++ runGuard/runguard.h
@@ -1,0 +1,29 @@
+#ifndef RUNGUARD_H
+#define RUNGUARD_H
+
+#include <QObject>
+#include <QSharedMemory>
+#include <QSystemSemaphore>
+#include <QCryptographicHash>
+
+class runGuard
+{
+public:
+    runGuard( const QString& key );
+    ~runGuard();
+
+    bool isAnotherRunning();
+    bool tryToRun();
+    void release();
+
+private:
+    const QString key;
+    const QString memLockKey;
+    const QString sharedmemKey;
+
+    QSharedMemory sharedMem;
+    QSystemSemaphore memLock;
+
+    Q_DISABLE_COPY( runGuard )};
+
+#endif // RUNGUARD_H